+++ packages/amplify-category-api/src/provider-utils/awscloudformation/legacy-add-resource.ts
@@ -1,3 +1,4 @@
+import { JSONUtilities } from 'amplify-cli-core';
 import { serviceMetadataFor } from './utils/dynamic-imports';
 import fs from 'fs-extra';
 import path from 'path';
@@ -29,24 +30,14 @@ export const legacyAddResource = async (serviceWalkthroughPromise: Promise<any>,
     copyCfnTemplate(context, category, answers, cfnFilename);
 
     const parameters = { ...answers };
-    const cfnParameters = {
-      authRoleName: {
-        Ref: 'AuthRoleName',
-      },
-      unauthRoleName: {
-        Ref: 'UnauthRoleName',
-      },
-    };
     const resourceDirPath = path.join(projectBackendDirPath, category, parameters.resourceName);
     fs.ensureDirSync(resourceDirPath);
 
     const parametersFilePath = path.join(resourceDirPath, parametersFileName);
-    let jsonString = JSON.stringify(parameters, null, 4);
-    fs.writeFileSync(parametersFilePath, jsonString, 'utf8');
+    JSONUtilities.writeJson(parametersFilePath, parameters);
 
     const cfnParametersFilePath = path.join(resourceDirPath, cfnParametersFilename);
-    jsonString = JSON.stringify(cfnParameters, null, 4);
-    fs.writeFileSync(cfnParametersFilePath, jsonString, 'utf8');
+    JSONUtilities.writeJson(cfnParametersFilePath, {});
   }
   context.amplify.updateamplifyMetaAfterResourceAdd(category, answers.resourceName, options);
   return answers.resourceName;