+++ packages/amplify-cli-core/src/index.ts
@@ -160,7 +160,7 @@ export type $TSCopyJob = any;
 
 // Temporary interface until Context refactor
 interface AmplifyToolkit {
-  confirmPrompt: (prompt: string, defaultValue?: boolean) => boolean;
+  confirmPrompt: (prompt: string, defaultValue?: boolean) => Promise<boolean>;
   constants: $TSAny;
   constructExeInfo: (context: $TSContext) => $TSAny;
   copyBatch: (context: $TSContext, jobs: $TSCopyJob[], props: object, force?: boolean, writeParams?: boolean | object) => $TSAny;
@@ -171,7 +171,7 @@ interface AmplifyToolkit {
   getPlugin: () => $TSAny;
   getCategoryPluginInfo: (context: $TSContext, category?: string, service?: string) => $TSAny;
   getAllCategoryPluginInfo: (context: $TSContext) => $TSAny;
-  getFrontendPlugins: () => $TSAny;
+  getFrontendPlugins: (context: $TSContext) => $TSAny;
   getEnvDetails: () => $TSAny;
   getEnvInfo: () => $TSAny;
   getProviderPlugins: (context: $TSContext) => $TSAny;
@@ -187,7 +187,7 @@ interface AmplifyToolkit {
   makeId: (n?: number) => string;
   openEditor: () => $TSAny;
   onCategoryOutputsChange: (context: $TSContext, currentAmplifyMeta: $TSMeta | undefined, amplifyMeta?: $TSMeta) => $TSAny;
-  pathManager: () => $TSAny;
+  pathManager: $TSAny;
   pressEnterToContinue: () => $TSAny;
   pushResources: (
     context: $TSContext,