+++ packages/amplify-cli/src/context-manager.ts
@@ -1,11 +1,11 @@
+import { stateManager } from 'amplify-cli-core';
+import { init } from './app-config';
+import { attachExtentions } from './context-extensions';
+import { NoUsageData, UsageData } from './domain/amplify-usageData';
+import { ProjectSettings } from './domain/amplify-usageData/UsageDataPayload';
 import { Context } from './domain/context';
 import { Input } from './domain/input';
 import { PluginPlatform } from './domain/plugin-platform';
-import { attachExtentions } from './context-extensions';
-import { init } from './app-config';
-import { UsageData, NoUsageData } from './domain/amplify-usageData';
-import { ProjectSettings } from './domain/amplify-usageData/UsageDataPayload';
-import { stateManager } from 'amplify-cli-core';
 
 export function constructContext(pluginPlatform: PluginPlatform, input: Input): Context {
   const context = new Context(pluginPlatform, input);
@@ -19,13 +19,12 @@ export async function attachUsageData(context: Context) {
   const usageTrackingEnabled = AMPLIFY_CLI_ENABLE_USAGE_DATA
     ? AMPLIFY_CLI_ENABLE_USAGE_DATA === 'true'
     : config.usageDataConfig.isUsageTrackingEnabled;
-  const accountId = await context.amplify.executeProviderUtils(context, 'awscloudformation', 'getAccountId');
   if (usageTrackingEnabled) {
     context.usageData = UsageData.Instance;
   } else {
     context.usageData = NoUsageData.Instance;
   }
-  context.usageData.init(config.usageDataConfig.installationUuid, getVersion(context), context.input, accountId, getProjectSettings());
+  context.usageData.init(config.usageDataConfig.installationUuid, getVersion(context), context.input, '', getProjectSettings());
 }
 
 const getVersion = (context: Context) => context.pluginPlatform.plugins.core[0].packageVersion;
@@ -34,9 +33,9 @@ const getProjectSettings = (): ProjectSettings => {
   const projectSettings: ProjectSettings = {};
   if (stateManager.projectConfigExists()) {
     const projectConfig = stateManager.getProjectConfig();
-    const frontend = projectConfig['frontend'];
+    const frontend = projectConfig.frontend;
     projectSettings.frontend = frontend;
-    projectSettings.framework = projectConfig?.[frontend].framework;
+    projectSettings.framework = projectConfig?.frontend?.framework;
   }
 
   if (stateManager.localEnvInfoExists()) {