+++ packages/amplify-provider-awscloudformation/src/iterative-deployment/deployment-manager.ts
@@ -7,6 +7,7 @@ import {
   DeploymentMachineStep,
   StateMachineHelperFunctions,
   createDeploymentMachine,
+  createRollbackDeploymentMachine,
   StateMachineError,
 } from './state-machine';
 import { IStackProgressPrinter, StackEventMonitor } from './stack-event-monitor';
@@ -131,11 +132,14 @@ export class DeploymentManager {
       const service = interpret(machine)
         .onTransition(async state => {
           if (state.changed) {
+            this.spinner.render();
             maxDeployed = Math.max(maxDeployed, state.context.currentIndex + 1);
             if (state.matches('idle')) {
               this.spinner.text = `Starting deployment`;
+            } else if (state.matches('deploy.waitForTablesToBeReady')) {
+              this.spinner.text = `Waiting for DynamoDB indices to be ready`;
             } else if (state.matches('deploy')) {
-              this.spinner.text = `Deploying stack (${maxDeployed} of ${state.context.stacks.length})`;
+              this.spinner.text = `Deploying (${maxDeployed} of ${state.context.stacks.length})`;
             } else if (state.matches('rollback')) {
               this.spinner.text = `Rolling back (${maxDeployed - state.context.currentIndex} of ${maxDeployed})`;
             } else if (state.matches('deployed')) {
@@ -155,6 +159,67 @@ export class DeploymentManager {
         })
         .start();
       service.send({ type: 'DEPLOY' });
+    });
+  };
+
+  public rollback = async (deploymentStateManager: IDeploymentStateManager): Promise<void> => {
+    this.deploymentStateManager = deploymentStateManager;
+    let currentStepIndex = this.deploymentStateManager.getStatus().currentStepIndex;
+    let maxDeployed = currentStepIndex + 1;
+
+    const deploymentTemplates = this.deployment.reduce<Set<string>>((acc, step) => {
+      acc.add(step.rollback.stackTemplatePath);
+      return acc;
+    }, new Set());
+    await Promise.all(Array.from(deploymentTemplates.values()).map(path => this.ensureTemplateExists(path)));
+    const fns: StateMachineHelperFunctions = {
+      preRollbackTableCheck: this.preRollbackTableCheck,
+      rollbackFn: this.rollBackStack,
+      tableReadyWaitFn: this.waitForIndices,
+      rollbackWaitFn: this.waitForDeployment,
+      stackEventPollFn: this.stackPollFn,
+      startRollbackFn: this.startRollBackFn,
+    };
+    const machine = createRollbackDeploymentMachine(
+      {
+        currentIndex: maxDeployed,
+        previousDeploymentIndex: currentStepIndex,
+        deploymentBucket: this.deploymentBucket,
+        region: this.region,
+        stacks: this.deployment,
+      },
+      fns,
+    );
+
+    return new Promise(async (resolve, reject) => {
+      const service = interpret(machine)
+        .onTransition(async state => {
+          if (state.changed) {
+            this.spinner.render();
+            if (state.matches('idle')) {
+              this.spinner.text = `Starting rollback`;
+            } else if (state.matches('preRollback')) {
+              this.spinner.text = `Waiting for previous deployment to finish`;
+            } else if (state.matches('rollback.waitForTablesToBeReady')) {
+              this.spinner.text = `Waiting for DynamoDB indices to be ready`;
+            } else if (state.matches('rollback')) {
+              this.spinner.text = `Rolling back (${maxDeployed - state.context.currentIndex} of ${maxDeployed})`;
+            } else if (state.matches('rolledBack')) {
+              this.spinner.succeed(`Rolled back succesfully`);
+            }
+          }
+
+          switch (state.value) {
+            case 'rolledBack':
+              return resolve();
+            case 'failed':
+              return reject(new DeploymentError(state.context.errors));
+            default:
+            // intentionally left blank as we don't care about intermediate states
+          }
+        })
+        .start();
+      service.send({ type: 'ROLLBACK' });
     });
   };
 
@@ -187,6 +252,22 @@ export class DeploymentManager {
     });
   };
 
+  public addRollbackStep = (rollbackStep: DeploymentOp): void => {
+    const rollbackStackTemplateUrl = getHttpUrl(rollbackStep.stackTemplatePathOrUrl, this.deploymentBucket);
+    const rollbackStackTemplatePath = getBucketKey(rollbackStep.stackTemplatePathOrUrl, this.deploymentBucket);
+
+    this.deployment.push({
+      deployment: null,
+      rollback: {
+        ...rollbackStep,
+        stackTemplatePath: rollbackStackTemplatePath,
+        stackTemplateUrl: rollbackStackTemplateUrl,
+        region: this.region,
+        clientRequestToken: rollbackStep.clientRequestToken ? `rollback-${rollbackStep.clientRequestToken}` : undefined,
+      },
+    });
+  };
+
   public setPrinter = (printer: IStackProgressPrinter) => {
     this.printer = printer;
   };
@@ -199,7 +280,7 @@ export class DeploymentManager {
     try {
       await this.deploymentStateManager?.startRollback();
       await this.deploymentStateManager?.startCurrentStep();
-    } catch {
+    } catch (err) {
       // ignore, rollback should not fail because updating the deployment status fails
     }
   };
@@ -238,15 +319,9 @@ export class DeploymentManager {
     return gsis ? gsis.every(idx => idx.IndexStatus === 'ACTIVE') : true;
   };
 
-  private waitForIndices = async (stackParams: DeploymentMachineOp) => {
-    if (stackParams.tableNames.length) {
-      console.log('\nWaiting for DynamoDB table indices to be ready');
-      // cfn is async to ddb gsi creation the api can return true before the gsi creation starts
-      await new Promise(res => setTimeout(res, 2000));
-    }
+  private waitForActiveTables = async (tables: string[]): Promise<void> => {
     const throttledGetTableStatus = throttle(this.getTableStatus, this.options.throttleDelay);
-
-    const waiters = stackParams.tableNames.map(name => {
+    const waiters = tables.map(name => {
       return new Promise(resolve => {
         let interval = setInterval(async () => {
           const areIndexesReady = await throttledGetTableStatus(name);
@@ -257,14 +332,25 @@ export class DeploymentManager {
         }, this.options.throttleDelay);
       });
     });
-
     await Promise.all(waiters);
+  };
+
+  private waitForIndices = async (stackParams: DeploymentMachineOp) => {
+    if (stackParams.tableNames.length) {
+      // cfn is async to ddb gsi creation the api can return true before the gsi creation starts
+      await new Promise(res => setTimeout(res, 2000));
+    }
+    await this.waitForActiveTables(stackParams.tableNames);
     try {
       await this.deploymentStateManager?.advanceStep();
     } catch {
       // deployment should not fail because saving status failed
     }
     return Promise.resolve();
+  };
+
+  private preRollbackTableCheck = async (stackParams: DeploymentMachineOp): Promise<void> => {
+    await this.waitForActiveTables(stackParams.tableNames);
   };
 
   private stackPollFn = (deploymentStep: DeploymentMachineOp): (() => void) => {
@@ -283,7 +369,7 @@ export class DeploymentManager {
 
   private doDeploy = async (currentStack: DeploymentMachineOp): Promise<void> => {
     try {
-      await this.deploymentStateManager?.startCurrentStep();
+      await this.deploymentStateManager?.startCurrentStep({ previousMetaKey: currentStack.previousMetaKey });
     } catch {
       // deployment should not fail because status could not be saved
     }