+++ packages/amplify-console-hosting/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-console-hosting@1.8.2...amplify-console-hosting@1.8.3) (2021-04-19)
+
+**Note:** Version bump only for package amplify-console-hosting
+
+
+
+
+
+## [1.8.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-console-hosting@1.8.1...amplify-console-hosting@1.8.2) (2021-04-14)
+
+**Note:** Version bump only for package amplify-console-hosting
+
+
+
+
+
 ## [1.8.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-console-hosting@1.7.5...amplify-console-hosting@1.8.1) (2021-04-09)
 
 **Note:** Version bump only for package amplify-console-hosting